Issue #SC-1511 feat: test case and code review fix
